Add additional scenario
